Fix performance-type-promotion-in-math-fn
- RandBeta() works now with double instead of float (this does not affect STEPWAT2 output at least on my machine)

warning: call to 'fabs' promotes float to double [performance-type-promotion-in-math-fn]
warning: call to 'exp' promotes float to double [performance-type-promotion-in-math-fn]
warning: call to 'log' promotes float to double [performance-type-promotion-in-math-fn]

+++ include/rands.h
@@ -48,7 +48,7 @@ void RandUniList(
     LOG_INFO *LogInfo
 );
 
-float RandBeta(float aa, float bb, sw_random_t *pcg_rng, LOG_INFO *LogInfo);
+double RandBeta(double aa, double bb, sw_random_t *pcg_rng, LOG_INFO *LogInfo);
 
 
 #ifdef __cplusplus
